use size_t instead of int for array size

+++ example/geometry/t8_example_geometries.cxx
@@ -100,7 +100,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -120,7 +120,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
@@ -160,7 +160,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -191,7 +191,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
@@ -224,7 +224,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -238,7 +238,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
@@ -280,7 +280,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -311,7 +311,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
@@ -351,7 +351,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -383,7 +383,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
@@ -424,7 +424,7 @@ public:
   void                t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
   {
     if (num_coords != 1)
@@ -440,7 +440,7 @@ public:
   void                t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                  t8_gloidx_t gtreeid,
                                                  const double *ref_coords,
-                                                 const int num_coords,
+                                                 const size_t num_coords,
                                                  double *jacobian) const
   {
     SC_ABORT_NOT_REACHED ();
