+++ src/t8_geometry/t8_geometry_implementations/t8_geometry_analytic.cxx
@@ -47,7 +47,7 @@ void
 t8_geometry_analytic::t8_geom_evaluate (t8_cmesh_t cmesh,
                                         t8_gloidx_t gtreeid,
                                         const double *ref_coords,
-                                        const int num_coords,
+                                        const size_t num_coords,
                                         double out_coords[3]) const
 /* *INDENT-ON* */
 
@@ -64,7 +64,7 @@ t8_geometry_analytic::t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                 t8_gloidx_t gtreeid,
                                                 const double
                                                 *ref_coords,
-                                                const int num_coords,
+                                                const size_t num_coords,
                                                 double *jacobian_out) const
 /* *INDENT-ON* */
