+++ src/t8_geometry/t8_geometry_implementations/t8_geometry_linear.cxx
@@ -45,7 +45,7 @@ void
 t8_geometry_linear::t8_geom_evaluate (t8_cmesh_t cmesh,
                                       t8_gloidx_t gtreeid,
                                       const double *ref_coords,
-                                      const int num_coords,
+                                      const size_t num_coords,
                                       double *out_coords) const
 {
   if (num_coords != 1)
@@ -60,7 +60,7 @@ t8_geometry_linear::t8_geom_evaluate_jacobian (t8_cmesh_t cmesh,
                                                t8_gloidx_t gtreeid,
                                                const double
                                                *ref_coords,
-                                               const int num_coords,
+                                               const size_t num_coords,
                                                double *jacobian) const
 {
   SC_ABORT ("Not implemented.");
